+++ apps/game/src/features/auth/index.ts
@@ -1,0 +1,11 @@
+// Auth Components
+export { default as Login } from './components/Login';
+export { default as Register } from './components/Register';
+export { default as TwoFactor } from './components/TwoFactor';
+export { default as Enable2FA } from './components/Enable2FA';
+
+// Auth Services
+export * from './services/auth.api';
+
+// Auth Store
+export * from './store/auth.store';
