+++ apps/game/src/features/game/index.ts
@@ -1,0 +1,5 @@
+// Game Components
+export { default as PongCanvas } from './components/PongCanvas';
+
+// Game Engine
+export * from './engine/engine';
