+++ apps/game/src/features/profile/index.ts
@@ -1,0 +1,2 @@
+// Profile Components
+export { default as Profile } from './components/Profile';
