+++ apps/game/src/localization/locales/index.ts
@@ -1,0 +1,3 @@
+export { en } from './en';
+export { ru } from './ru';
+export { zh } from './zh';
