+++ apps/game/src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App";
-import { TranslationsProvider } from "./translations";
+import App from "./app/App";
+import { TranslationsProvider } from "./localization";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
