--- apps/game/src/profile/index.ts
@@ -1,1 +1,0 @@
-export { default as Profile } from './Profile';
