+++ apps/game/src/shared/components/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as Banner } from './Banner';
