+++ apps/game/src/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './components';
+export * from './store';
