+++ apps/game/src/shared/store/index.ts
@@ -1,0 +1,1 @@
+export * from './ui.store';
