+++ apps/game/src/shared/store/ui.store.ts
